requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
python-dotenv==1.0.0